fix: vless REALITY fp param, transport layer, DNS+rules, fetcher retry
BUG FIXES:
- generator: vless REALITY URLs now include fp= (previously lost)
- generator: vless REALITY Clash proxies now include client-fingerprint
- generator: vmess/http transport type now serialized to URL + Clash
- generator: vless/http transport type now serialized to URL + Clash
- generator: trojan/http transport type now serialized to URL + Clash
- generator: trojan ws-opts now includes headers field
- parser: URL regex https? scheme matching fixed

FEATURES:
- generator: Clash YAML now has full DNS config, mixed-port, CN split rules
- generator: sing-box JSON now has DNS, mixed inbound, rule_set CN bypass
- fetcher: _fetch_one now retries up to 2 times with 1s/2s backoff
- fetcher: handles 429 rate-limit with Retry-After header
- tester: latency_ms changed from int to float for precision
- main: per-protocol pass/fail stats in stats.json
- main: node tags show float latency (e.g. [123.4ms])

+++ core/tester.py
@@ -45,7 +45,7 @@ MIN_LATENCY_MS = 30  # 低于此值视为同机房假节点
 class TestResult:
     node: Node
     success: bool = False
-    latency_ms: int = 0
+    latency_ms: float = 0.0
     error: str = ""
